src/requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
idna==3.4
joblib==1.3.2
nltk==3.8.1
oauthlib==3.2.2
python-dotenv==1.0.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
tqdm==4.66.1
tweepy==4.14.0
urllib3==2.0.6
